+++ JLinkExamples/AddTwo.java
@@ -1,0 +1,3 @@
+public class AddTwo {
+	public static int addtwo(int i, int j) {return i + j;}
+}